+++ passport/strategies/user/strategy-local.js
@@ -1,11 +1,10 @@
 /**
  * Created by championswimmer on 07/05/17.
  */
+const Raven = require('raven')
 const LocalStrategy = require('passport-local').Strategy;
 const models = require('../../../db/models').models;
 
-const secrets = require('../../../secrets.json');
-const config = require('../../../config');
 const passutils = require('../../../utils/password');
 
 
@@ -36,6 +35,6 @@ module.exports = new LocalStrategy(function (username, password, cb) {
                 return cb(err, false, {message: err})
             });
 
-    });
+    }).catch((err) => Raven.captureException(err))
 
 });