+++ passport/strategies/user/strategy-twitter.js
@@ -1,12 +1,13 @@
 /**
  * Created by championswimmer on 07/05/17.
  */
+const Raven = require('raven')
 const TwitterStrategy = require('passport-twitter-email').Strategy;
 
 const models = require('../../../db/models').models;
 
-const secrets = require('../../../secrets.json');
 const config = require('../../../config');
+const secrets = config.SECRETS;
 const passutils = require('../../../utils/password');
 
 /**
@@ -35,7 +36,7 @@ module.exports = new TwitterStrategy({
             return models.User.findById(oldUser.id)
         }).then(function (user) {
             return cb(null, user.get())
-        })
+        }).catch((err) => Raven.captureException(err))
     } else {
 
         models.User.count({where: {username: profileJson.screen_name}})
@@ -49,7 +50,7 @@ module.exports = new TwitterStrategy({
                         tokenSecret: tokenSecret,
                         username: profileJson.screen_name,
                         user: {
-                            username: existCount == 0 ? profileJson.screen_name : profileJson.screen_name + "-t",
+                            username: existCount === 0 ? profileJson.screen_name : profileJson.screen_name + "-t",
                             firstname: profileJson.name.split(' ')[0],
                             lastname: profileJson.name.split(' ').pop(),
                             email: profileJson.email || undefined,
@@ -63,7 +64,7 @@ module.exports = new TwitterStrategy({
                 return cb(null, false);
             }
             return cb(null, userTwitter.user.get())
-        })
+        }).catch((err) => Raven.captureException(err))
 
 
     }