+++ routers/api/index.js
@@ -5,6 +5,14 @@
  */
 const router = require('express').Router();
 
+router.use((req, res, next) => {
+    res.set('Access-Control-Allow-Headers', 'X-Requested-With,content-type,Authorization')
+    res.set('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, PATCH, DELETE')
+    res.set('Access-Control-Allow-Origin', req.get('origin'))
+    res.set('Access-Control-Allow-Credentials', true)
+    next()
+})
+
 router.use('/users', require('./users'));
 router.use('/clients', require('./clients'));