+++ routers/api/users.js
@@ -10,7 +10,8 @@ const models = require('../../db/models').models;
 
 
 router.get('/me',
-    passport.authenticate('bearer', {session: false}),
+    // Frontend clients can use this API via session (using the '.codingblocks.com' cookie)
+    passport.authenticate(['bearer', 'session']),
     function(req,res) {
 
         if (req.user) {
@@ -46,6 +47,28 @@ router.get('/me',
 
     });
 
+router.get('/me/logout',
+  passport.authenticate('bearer', {session: false}),
+  function (req, res) {
+    if (req.user) {
+      models.AuthToken.destroy({
+        where: {
+          token: req.header('Authorization').split(' ')[1]
+        }
+      }).then(function () {
+        res.status(202).send({
+          'user_id': req.user.id,
+          'logout': 'success'
+        })
+      }).catch(function(err) {
+        res.status(501).send(err)
+      })
+    } else {
+      res.status(403).send("Unauthorized")
+    }
+  }
+)
+
 router.get('/:id',
     passport.authenticate('bearer', {session: false}),
     function (req, res) {
@@ -53,6 +76,8 @@ router.get('/:id',
             return res.send(req.user)
         }
         models.User.findOne({
+            // Public API should expose only id, username and photo URL of users
+            attributes: ['id', 'username', 'photo'],
             where: {id: req.params.id}
         }).then(function (user) {
             if (!user) {