+++ routers/login/github.js
@@ -6,11 +6,26 @@
 const router = require('express').Router();
 const passport = require('../../passport/passporthandler');
 
+function authnOrAuthzGithub(req, res, next) {
+  if (!req.isAuthenticated()) {
+    if(config.DEBUG) console.log("Authn Github = = = = = ");
+    passport.authenticate('github', {
+      failureRedirect: '/login',
+      successReturnToOrRedirect: '/users/me'
+    })(req, res, next);
+  } else {
+    if(config.DEBUG) console.log("Authz Github = = = = = = ");
+    passport.authorize('github', {
+      //TODO: Add failure flash 
+      failureRedirect: '/user/me'
+    })(req, res, next);
+  }
+}
+
 router.get('/', passport.authenticate('github'));
 
-router.get('/callback', passport.authenticate('github', {
-    failureRedirect: '/login',
-    successReturnToOrRedirect: '/users/me'
-}));
+router.get('/callback', authnOrAuthzGithub, function (req, res, next) {
+    res.redirect('/users/me')
+});
 
 module.exports = router;