+++ routers/login/twitter.js
@@ -14,9 +14,9 @@ function authnOrAuthzTwitter(req, res, next) {
         })(req, res, next);
     } else {
         if(config.DEBUG) console.log("Authz Twitter = = = = = = ");
-        passport.authenticate('twitter', {
-            failureRedirect: '/login',
-            successReturnToOrRedirect: '/users/me'
+        passport.authorize('twitter', {
+            //TODO: Add failure flash (could not connect your account to ...)
+            failureRedirect: '/users/me'
         })(req, res, next);
     }
 }
@@ -24,6 +24,8 @@ function authnOrAuthzTwitter(req, res, next) {
 
 router.get('/', passport.authenticate('twitter'));
 
-router.get('/callback', authnOrAuthzTwitter);
+router.get('/callback', authnOrAuthzTwitter, function (req, res, next) {
+    res.redirect('/users/me');
+});
 
 module.exports = router;