+++ routers/signup.js
@@ -6,29 +6,57 @@
 const router = require('express').Router();
 const models = require('../db/models').models;
 const passutils = require('../utils/password');
+const makeGaEvent = require('../utils/ga').makeGaEvent
 
-router.post('/', function (req, res) {
-    passutils.pass2hash(req.body.password)
-        .then(function (passhash) {
-            models.UserLocal.create({
-                user: {
-                    username: req.body.username,
-                    firstname: req.body.firstname,
-                    lastname: req.body.lastname,
-                    email: req.body.email
-                },
-                password: passhash
-            }, {
-                include: [models.User]
-            }).then(function (user) {
-                res.redirect('/login');
-            })
 
-        })
-        .catch (function (err) {
-            // Could not register user
-            //TODO: Handle this case (use flash message maybe ? )
-        });
+router.post('/', makeGaEvent('submit', 'form', 'signup'), function (req, res) {
+
+  if(req.body.username.trim() === '') {
+    req.flash('error', 'Username cannot be empty');
+    return res.redirect('/signup')
+  }
+  if((req.body.firstname.trim() === '') || (req.body.lastname.trim() === '')) {
+    req.flash('error', 'Firstname and/or Lastname cannot be empty');
+    return res.redirect('/signup')
+  }
+  if(req.body.email.trim() === '') {
+    req.flash('error', 'Email cannot be empty');
+    return res.redirect('/signup')
+  }
+  if((req.body.password.trim() === '') || req.body.password.length < 5) {
+    req.flash('error', 'Password too weak. Use 5 characters at least.');
+    return res.redirect('/signup')
+  }
+
+    models.User.findOne({where:{username:req.body.username}})
+    .then((user)=> {
+      if(user)
+      {
+          req.flash('error', 'Username already exist\'s. Please try again.');
+          return res.redirect('/signup')
+      }
+      passutils.pass2hash(req.body.password)
+          .then(function (passhash) {
+              models.UserLocal.create({
+                  user: {
+                      username: req.body.username,
+                      firstname: req.body.firstname,
+                      lastname: req.body.lastname,
+                      email: req.body.email
+                  },
+                  password: passhash
+              }, {
+                  include: [models.User]
+              }).then(function (user) {
+                  res.redirect('/login');
+              })
+          })
+    })
+    .catch (function (err) {
+        // Could not register user
+         req.flash('error', 'Unsuccessful registration. Please try again.');
+        return res.redirect('/signup')
+    });
 });
 
 module.exports = router;