+++ server.js
@@ -1,27 +1,36 @@
 /**
  * Created by championswimmer on 08/03/17.
  */
+require('newrelic')
 const express = require('express')
     , bodyParser = require('body-parser')
     , session = require('express-session')
     , passport = require('./passport/passporthandler')
     , path = require('path')
-    , cookieParser = require('cookie-parser')
     , exphbs = require('express-hbs')
-    , expressGa = require('express-ga-middleware');
+    , expressGa = require('express-ga-middleware')
+    , flash = require('express-flash')
+    , Raven = require('raven')
+    , debug = require('debug')('oneauth:server')
 
-const secrets = require('./secrets.json')
-    , config = require('./config')
+const config = require('./config')
+    , secrets = config.SECRETS
     , loginrouter = require('./routers/login')
     , connectrouter = require('./routers/connect')
+    , disconnectrouter = require('./routers/disconnect')
     , logoutrouter = require('./routers/logoutrouter')
     , signuprouter = require('./routers/signup')
     , apirouter = require('./routers/api')
     , oauthrouter = require('./routers/oauthrouter')
     , pagerouter = require('./routers/pagerouter')
-    , statusrouter = require('./routers/statusrouter');
+    , statusrouter = require('./routers/statusrouter')
+    , {expresstracer, datadogRouter} = require('./utils/ddtracer');
 
 const app = express();
+
+// ============== START DATADOG
+app.use(expresstracer)
+// ================= END DATADOG
 const redirectToHome = function (req, res, next) {
 
     if (req.path == '/') {
@@ -32,6 +41,11 @@ const redirectToHome = function (req, res, next) {
 
 };
 
+// ====================== START SENTRY
+Raven.config(secrets.SENTRY_DSN).install()
+app.use(Raven.requestHandler())
+// ====================== END SENTRY
+
 app.engine('hbs', exphbs.express4({
     partialsDir: path.join(__dirname, 'views/partials'),
     layoutsDir: path.join(__dirname, 'views/layouts'),
@@ -41,21 +55,26 @@ app.set('views', path.join(__dirname, 'views'));
 app.set("view engine", "hbs");
 
 app.use(express.static(path.join(__dirname, 'public_static')));
-app.use(cookieParser(secrets.EXPRESS_SESSION_SECRET));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(session({
     secret: secrets.EXPRESS_SESSION_SECRET,
     resave: false,
     saveUninitialized: false,
-    name: 'oneauth'
+    name: 'oneauth',
+    cookie: {
+        domain: config.COOKIE_DOMAIN
+    }
 }));
+app.use(flash());
 app.use(passport.initialize());
 app.use(passport.session());
 app.use(redirectToHome);
 app.use(expressGa('UA-83327907-7'));
+app.use(datadogRouter)
 app.use('/login', loginrouter);
 app.use('/connect', connectrouter);
+app.use('/disconnect', disconnectrouter);
 app.use('/logout', logoutrouter);
 app.use('/signup', signuprouter);
 app.use('/api', apirouter);
@@ -63,9 +82,8 @@ app.use('/oauth', oauthrouter);
 app.use('/', pagerouter);
 app.use('/status',statusrouter);
 
-app.listen(3838, function () {
-    console.log("Listening on " + config.SERVER_URL );
+app.use(Raven.errorHandler());
+
+app.listen(process.env.PORT||3838, function () {
+    debug("Listening on " + config.SERVER_URL );
 });
-
-
-